🔨 Refactor: 핑퐁 게임 수정
