+++ src/app/pingpong/Pingpong.module.css
@@ -1,0 +1,22 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: 100%;
+  background-color: black;
+  color: white;
+}
+
+.recordBoard {
+  display: flex;
+  justify-content: space-around;
+  width: 800px;
+}
+
+.gameCanvas {
+  border-radius: 10px;
+  border: 2px solid #4caf50;
+  background-color: black;
+}
